[FIXED] handle int max

+++ cpp08/ex01/Span.cpp
@@ -31,7 +31,7 @@ void	Span::addNumber(int number)
 	this->_numbers.push_back(number);
 }
 
-int	Span::longestSpan()
+unsigned int	Span::longestSpan()
 {
 	if (this->_numbers.size() < 2)
 	{
@@ -40,10 +40,10 @@ int	Span::longestSpan()
 	int minVal = *std::min_element(this->_numbers.begin(), this->_numbers.end());
 	int maxVal = *std::max_element(this->_numbers.begin(), this->_numbers.end());
 
-	return maxVal - minVal;
+	return static_cast<unsigned int>(static_cast<long long>(maxVal) - static_cast<long long>(minVal));
 }
 
-int	Span::shortestSpan()
+unsigned int	Span::shortestSpan()
 {
 	if (this->_numbers.size() < 2)
 	{
@@ -53,7 +53,8 @@ int	Span::shortestSpan()
 	std::vector<int> sorted = this->_numbers;
 	std::sort(sorted.begin(), sorted.end());
 
-	int minSpan = sorted[1] - sorted[0];
+	// secure (minimum)
+	long long minSpan = static_cast<long long>(sorted[1]) - static_cast<long long>(sorted[0]);
 	for (size_t i = 1; i < sorted.size() - 1; ++i)
 	{
 		int diff = sorted[i + 1] - sorted[i];
@@ -62,5 +63,5 @@ int	Span::shortestSpan()
 			minSpan = diff;
 		}
 	}
-	return minSpan;
+	return static_cast<unsigned int>(minSpan);
 }
